upgrade OpenSBI to 0.9
The resulting binary is larger so we have to change the memory
mapping in low_load.

The current version of OpenSBI is 1.2 but we cannot easily upgrade
past 0.9 because the thead platform has been merged into generic
so we would have to add the uart driver in a different way.